C backend: Fixed handling of outer refs
For now, outer refs are always pointers to the outer instance.  The interpreter
uses an optimization for standard value types such as i32 to treat them as
values.

This optimization will need additions to FUIR, but for now the only goal is to
get a working and simple C backend, optimizations will be added later.