Loops are now all converted to tail recursive features
The Parser directly produces an AST that no longer contains any loops.  Please
check the long comment at the beginning of Loop.java for details on what this
code looks liek.

The result is a significant simplification of the later phases, code is reduced
by almost 500 lines and backend support for loops is no longer needed.

The C backend can translate loops now even though this patch does not touch the
C backend.

Since the back ends do not support tail recursion yet, the stack sizes in some
scripts and Makefiles had to be increased.