Tools: Replace Errors.showStatistics with Errors.showAndExit
Errors.showAndExit checks if errors had occured and if so, shows their number
and exits with error code 1.

Made interpreter and C backend code behave equal when errors found in up to
findAllClazzes.

Added a call to showAndExit after execution of a Fuzion tool to always report
the error count and exit with an error in case any problem occured.