Parser/lib: Make parentheses optional in most calls
Inspired by today's FOSDEM talks on Raku, I have decided to experiment with
making the parentheses on calls optional.

For this to work, however, I had to introduce a new keyword 'then' to separate
the condition in a single-line 'if'-statement from the following block.

Also, it is no longer possible to declare an uninitialized field using the
syntax

   f type

This syntax is was hardly ever used in more recent examples, so it is ok to
make this a bit more complex. An uninitilaized field is no declared as

   f type = ?

and I think I might completely remove this unless I find a really convincing
use case.