IR: Centralized creation of result clazz of a feature to Clazz.actualResultClazz
This avoids repetition of special handling for the result of outer references,
which were incorrect in some cases.  The code still needs to tested and maybe
fixed for outer refs that are inherited.

This change avoids the generation of fictional outer clazzes like
hasInterval<i32>, which caused the C backend to crash.