lib: make use of boxing when assigning to list, which is a choice type
cons cells do not need to be marked as 'ref' explicitly, it is sufficient that
list is a choice of nil or Cons, which is a ref already.

Unfortunately, this changes unmasks a problem in the IR when generating clazzes
for actual value instances of hasInterval, so 'one' cannot be used directly
until this is fixed.