lib/C backend: Remove C compiler warning when compiling ref i32.
The problem was that ref i32 is a ref, while i32.thiz is an i32, so returning
Current resulted in wrong C code.

Now, i32/u32/i64/u64.thiz return this.val, which is always a value type, and the
C backend now has a special handling for accessing the 'val' field in a scalar, such that the warning is gone and the code should even work (not tested...:-).