FUIR/C backend: Fix dynamic calls, many target clazzes were missing
Since FUIR.callCalledClazzes() contained only the inner clazzes, different
targets that share the same inner clazzes in a call where not taken into account
in the code for a dynamic call.

Now, FUIR.callCalledClazzes() returns a list of pairs target, inner that gives
all possible targets in a call together with the corresponding inner clazzes.