Parser: Extended OpExpr parser to take care of white space
So far, only the Parser for calls ensured that 'a -b' is parsed as not parsed as
an infix operation 'a.infix -(b)', but as a call 'a(-b)'.  This patch extends
this to general operator expressions. I.e. in the following expressions

  a- * b
  a * -b

The '+' will be given higher precedence than the '*' due to the usage of white
space, so these will be parsed as

  (a-) * b
  a * (-b)

and not as

  a - (*b)
  (a*) -b