IR/FUIR/C backend: Add support for array of arbitrary element types
For this, IR and FUIR have to provide a means to access the actual generics of a
clazz.

For easier lookup, the intrinsics of Array have been changed to be generics
themselves such that there is no need to search for the actual generic in the
ouuter clazz's inheritance hierarchy.