IR: Handle outer of type Object as always instantiated.
Once Clazz.normalize() is implemented better, a clazz C has to be considered
instantiated if there is any clazz D that normalize() would replace by C if it
occurs as an outer clazz.

This fix the C backend for test design/examples/red_field.fz, which creats a
boxed instance of unit.