Parser: Support '$<id>' and '{<expr>}' to embed expressions into strings
This syntactic sugar is implemented mostly in the lexer, with some support from
the parser.

There are two ways to access the closure within a constant string: Via '$' or
via '{}'.  A '$' encountered in a constant string will result in the following
token that must start with a letter or digit being parsed as an expr.
Similarly, a '{' encountered in a constant string will result in the following
tokens to be parsed as an expr until a matching '}' is parsed.

In both cases, the parsed expression will be concatenated using 'infix +' with
the beginning and the end of the string.

Examples:

  x := 42
  say "x is $x."                         # 'x is 42.'
  say "the answer is $42."               # 'the answer is 42.'
  a := 3
  b := 4
  say "sum is {a+b} product is {a*b}"    # 'sum is 7 product is 12'

An '{' in a constant string serves logically like a closing '"', finishing a
partial string.  The following code will be parsed like an expression without
any restrictions except for the final '}' that starts the next partial string.

'{'and '}' may be nested, i.e., it is possible to write

  say "max(a,b) is {if (a>b) a else b}." # max(a,b) is 4.
  say "min(a,b) is {if (a<b) a else b}." # min(a,b) is 3.

It is even possible to nest strings into expressions embedded using '{' and '}'
and then, again, embed expressions into these, e.g.,

  say "Nested: { "-{ a + b }-" * 3 }"    # 'Nested: -7--7--7-'

Here

  a + b

is embedded in the inner string

  "-{ a + b }-" ,

which is then repeated 3 times using

  "-{ a + b }-" * 3

which is finally embedded into

  "Nested: { "-{ a + b }-" * 3 }"