ast: Enable generic argument inference for match cases
In a 'match' statement with a case that includes generic arguments, the generic
arguments can be omitted iff they are not required to uniquely identify that
case.  In particular, if the choice type does not have different type parameters
that have the same base class (such as 'array<bool>' and 'array<Object>'), the
generics can be omitted.

This reduces the amount of typing, in particular when matching Cons-cells in
lists: Instead of 'Cons<A, list<A>>', it is now sufficient to match against
'Cons'.