Parser/FE: add support for array initialization
So far only in the parser: Arrays can now be intialized using '[a, b, c]' or
'[e; f; g]' where the array elements are expressions.

The array type is either inferred from the union of the types of the element
expressions or propagated back from the target the array is assigned to.  Boxing
and Tagging for array elements is automatic.