C backend: Fill unused bytes in conststring with '\000' instead of '.'
These unused bytes are there since conststring still is an array<i32>, should be
come an array<u8>.  The garbage '.' stored in there resulted in a difference
when comparing byte-by-byte with a string created dynamically.

This fixes tests/strings and tests/visibility when usning C backend, bringing
the number of failing tests for the C backend down to 10 (5 more than the
interpreter).