FUIR/C backend: Support for precondition checks in inheritance calls
This is a bit of a hack: Calls to parent features in inherits clauses are
inlined, so the backend does not need to take care.  Only the precondition are
executed explicitly as a call with FUIR.callPreconditionOnly() being true.

This change also required some cleanup in Clazz.called/instantiated and the
calls to these for internally called and instantiated features.