C backend: Stop code generation after a call with a void result
A call with a void result is guaranteed never to return.  Any code generated
after such a call may cause problems to our C backend or the C compiler.

A null value on the stack is used to transfer the information that a void
result was encountered and code generation must stop.