lib/numeric&wrappingInteger: Add unit types for integer classes
New unit types i32s, i64s, u32s, u64s provide features requires for numeric and
wrappingInteger that do not require an instance of the respective number, such
as constants zero, one, min, max and monoids defined for numeric types.

numerics.fz is a new unit type the provides generic impelmentation of monoids
for numeric types.

wrappingIntegers.fz is a new unit type that defines the requires featuers min
and max that do not depend on an instance of the numeric type and are used by
wrappingInteger.