lib/string: Avoid quadratic runtime when using concatenated strings
Reading one byte from a string that is the result of concatenating string
a0+a1+a2+..+a<n> required time in O(n), such that printing the string was in
O(n²), which is unacceptable.

Now, repeated concatenation collects the partial strings in a growing array,
resulting in linear performance both for creation and for iterating the bytes of
a string.