lib: renamed streamable as List
List is a reference type used as parent for data structures that can be seen as
a list of values (arrays, sets, etc.).  From any such data structure, a lazy
'list' or 'stream' can be created.

In contrast, a 'list' is a lazy lisp-style list that is either 'nil' or a 'Cons'
cell consisting of a 'head' element and a 'tail', which is also a 'list'. 'list'
inherits from 'List', while 'List' can be converted to 'list' via 'List.asList'
(which should become implicit once type conversions are supported), such that
both are interchangeable.

To avoid filename clashes on case-insensitive systems, renamed list.fz as
list0.fz.