FUIR: Fix precondition failure when compiling inherits clause with arguments
The problem is a call to Clazz.lookup to find the argument fields of the called
feature at a stage after the clazzes have been created.

This fix creates the corresponding clazzes during the findClazzes stage and
reuses the result in FUIR.

This fixes compilation of the idiom examples idiom100ex.fz, idiom124ex.fz, and
some more.