ir: Reduce number of clazzes by normalization of outer clazz
Two clazzes for the same feature that only differ by their outer clazz being
different ref types are now fused into the same clazz by normizing the outer
clazz.  This results in a signficicant reduction in the number of runtime
clazzes generated.

Normalization happens only for outer clazzes that are refs and that do not have
inner clazzes that make use of the outer ref to the outer clazz.

This required some changes in FUIR and the C backend as well:

 - FUIR.clazzNeedsCode() is a replacement of FUIR.clazzIsCalled().  This is true
   for clazzes that require code generated by the backend.

 - an assignment now unboxes a value if it comes from an outer ref to a value
   type.