FE/Type: Fix dependsOnGenerics: false if actual generics dont depend on generics
With the original code, dependsOnGenerics was true for any type that had actual
generics, even if these actual generics did not depend on any type parameters.

Now, dependsOnGenerics() is true only for type parameters or, recursively, if
any actual type parameters or the outer type depend on genenerics.

This fixes a major performance issue in the front end and middle end and
Type.actualType() uses dependsOnGenerics() to decide if a type needs to be
checked for type parameters or not.

I measured a performance gain of this patch between 38% (for a HelloWorld) up to
61% for a larger example.