C backend: Fix struct declaration order
The struct of the type referred to by an outer ref does not need to be declared
before the inner type's struct since outer refs are references, even if the
outer clazz is a value type.

This fixes C compilation problems in case the outer clazz contains fields whose
type is the inner clazz and the outer clazz' struct is declared first due to
the outer ref.