Parser: New call syntax 'f a b' as alternative to 'f(a,b)'.
This patch introduces Haskell-style call syntax with space separated arguments
and left binding, so the call

  f a b (g c d)

is equivalent to

  f(a,b,g(c,d))

Additionally, infix, prefix and postfix operators now bind stronger than a call
if there is no white space (or comment) between the operator and its operand,
and weaker than a call if there is white space.  This means

  f x+1

is equivalent to

  f(x+1)

while

  f x + 1

is equivalent to

  f(x)+1

This makes the semantics depend on white space in a way that might be
surprising.  This change is therefore considered for experiments only, if this
turns out to be too confusing, the grammar should go back to not make operator
precedence depend on spaces around operators.