FE, IR, Interpreter: Several fixes related to normalized outer clazzes
This fixes several tests that failed when using C backend and some that failed
for the interpret.

Fixes idion134ex.fz that did not use dynamic binding properly.

Dynamic binding was disabled for calls with a value target type, which is
basically correct except if the target is on outer instance it might at runtime
have been turned into a ref type, so this check was moved from the front end
(Call.java) to the IR.

This required special handling in the interpreter to access the 'val' field of
ref versions of basic integer types i32, u32, etc.

The normalization of outer clazzes resulted in wrong code for some compiled
tests, in particular the detection of what clazzes are instantiated depending on
ther outer clazzes being instantiated was improved.

Finally, field result type inference now uses the actual clazz of the result
expression instead of trying to derive this clazz from the type determined by
the front end.