FUIR/C backend: Implement match for choice of only ref values
A match for a choice that consists of different kinds of ref values is
implemented by checking the clazzes of the actual instances instead of checking
the tag value.  The clazzes might be heirs of the static clazz in the choice
type, so heirs have to be taken into account as well.