ir/Clazz: Separate determination of fields from clazz layout, fix #44
The problem was that isUnitType() requires the set of fields, and this is called
before class layout when it is checked which choice clazzes require a tag.
Consequently, some tag may be missing in the IR resulting in the
NullPointerExcepion seen in #44.

This fix untangles the code: Clazz.fields() determines the set of fields in a
clazz, while Clazz.layout() uses Clazz.fields to make sure clazz layout would
work (i.e., a clazz is not recursive).  The actual layout is done in the
backend.