fzjava,be/Interpreter,excamples/webserver: Allow methods to throw exceptions
Java method that declare to throw exceptions will see their result wrapped in an
instance of 'outcome' that will contain either the result of an 'error' created
form the Java exception that was thrown.  This is true even for Java methods
that result in (Java's) 'void', which, if they declare to throw an exception,
will now result in 'outcome<unit>'.

The webserver example was extended to handle these exceptions and with a comment
explaining what syntactic sugar is planned to reduce the boilerplate for this.