FE/Errors: Provide a solution in case 'result' is used when no return type is given
This is a first step to improve the behaviour in #44 by providing a hint on how
to get an automatically declared result field.  This does not help in the
example of #44, though, since 'result' should be visibly in the postcondition.

It helps in this example:

  testResult
    pre
      result != 3
    post
      result = 3
  is
    set result := 3
    if result = 3
      say "result is 3"

which now results in

 > ./build/bin/fz testResult

./testResult.fz:7:14: error 1: Could not find target field 'result' in assignment
  set result := 3
-------------^
Field not found: 'result'
Within feature: 'testResult'
For assignment: 'result = 3'
To solve this, make sure you declare a return type in the surrounding feature such that the 'result' field will be declared automatically.

./testResult.fz:8:6: error 2: Could not find called feature
  if result = 3
-----^
Feature not found: result (no arguments)
Target feature: 'testResult'
In call: 'result'
To solve this, make sure you declare a return type in the surrounding feature such that the 'result' field will be declared automatically.

./testResult.fz:3:5: error 3: Could not find called feature
    result != 3
----^
Feature not found: result (no arguments)
Target feature: 'testResult'
In call: 'result'
To solve this, make sure you declare a return type in the surrounding feature such that the 'result' field will be declared automatically.

./testResult.fz:5:5: error 4: Could not find called feature
    result = 3
----^
Feature not found: result (no arguments)
Target feature: 'testResult'
In call: 'result'
To solve this, make sure you declare a return type in the surrounding feature such that the 'result' field will be declared automatically.

4 errors.