lib, be: string.utf8 now is List<u8> not List<i32>
A string is a sequence of unsigned bytes of type u8 that give the utf8 encoding
of the string.

This requires changes on all code iterating the utf8 bytes of a string and on
all features that inherit from string such as conststring.

Also required changes in the C backend (for conststring) and in the interpreter
(for conststring and support for array of integers that are not i32).