lib: Remove no longer needed explicit type conversion from constants
'10.as_u8' can now be replaced by '10' if this value is a assigned to a field
tof type 'u8', so removed this cast.

Where explicit types are still required, replaced '0xff.as_u8' by 'u8 0xff'.
This syntax works since it is a call to 'u8(val u8)', so '0xff' is assigned to
'val' which is of type 'u8', which is propagated to the constant.