Parser: Extend the Lexer to support float literals
Numeric literals are of the form

  {digits}+[.{digits}][["+"|"-"]("E"|"P"){digits}.

The digits before the '.' or the digits of the exponent may have a radix prefix
'0b', '0o', '0d' or '0x' and may use '_' to group digits into groups of equal
sizes.  The digits after '.'  may not have a radix prefix and may not contain
'_'.

Here are some examples

  1234
  1234.5678
  1234.5678E9
  1234.5678P9
  0b0.11P0b100       (eq. to 3)
  1_000_000.000001
  0xf_ade_0ffE-23
  0xBEEFP0xBABE

Negative examples are

  0abc          -- hex number not allowed
  0x123E0x45    -- the 'E' is part of the hex number, not an exponent
  1_234.567_890 -- '_' not allowed after '.'