FE: Enfore result rexpression in constructor to be assignable to 'unit', fix #47
Declaring a constructor instead of a routine with a result is a common mistake
in Fuzion, so silently ignoring the result was not a good idea.

Now, when one tries to produce a non-unit result in a constructor, an error will
be produces that suggests to either turn the construtor into a routine with a
compatible result type or to explicitly return unit.