FE: Fix assignment to inherited field in outer feature, fix #35
The problem was that the access to the outer instance did not consider that
outer might be a heir of assignedField.outer().

Now, the code to access the target instance of an assignment is the same as the
code to access the implicit target of an unqualified call, the code was moved
from Call to Feature.java and is now used in Assign.java as well.