FUIR/C backend: Fix code generation for assignment to outer field, fix #80
In case the target clazz of an assignment to a field is a reference we might
need dynamic binding when writing the field.

This patch unifies the code for dynamic binding for a call to a feature and
dynamic binding for writing a field to avoid code duplication.

In FUIR, APIs for Calls (call*) and Assignments (assign*) have been unified as
access* to avoid duplication.