IR: Fix wrong cashing of Clazz.lookup for open generic
The problem is that lookup results are cached for empty actual generics. In case
of open generics, however, there might be emty generic lists that are cached and
non-empty lists that are checked.  So we must use the cached value only if the
actual generics are also empty.

This fixes failure of tests/tuples.