Parser: New syntax for lambdas: 'x,y -> x*y' or '(x,y) -> x*y'
This new syntax for lambdas uses type inference from the target the expression
is assigned to to avoid the need to define the types. The input arguments may be
enclosed in parentheses, i.e.,

  (x,y) -> x*y

is equivalent to

  x,y -> x*y

An empty list of arguments can be defined using '()', e.g.,

  () -> 42

Type inferencing from a lambda expression as in

  inc := x -> x+1   # error!
  f => x -> x+1     # error!

will result in an error since no type if given. Instead, we need an explicit
type

  f i32 -> i32 is x -> x+1

or a generic type

  f<T: numeric<T>> T -> T is x -> x+1