FE+IR: Fix handling of Expr.typeOrNull end Expr.type()
typeOrNull() may be called during type inferencing and may return null, while
type() is guaranteed not to return null, but will cause an error if the type is
not known yet.

This patch uses type() in the later phases instead of typeOrNull() and adds
handling of null results for typeOrNull().