lib: for integer types, us 'thiz' instead of '<xyz>.this'
Since '<xyz>.this' cannot be assigned to '<xyz>' in the general case since it
might be a value sub-type of '<xyz>', an upcoming patch will create an error for
such assignments.  Using 'thiz' for the integer types makes sure this error does
not show up.

Eventually, Fuzion should support an Eiffel-style 'like this' (or 'like outer')
type that could then be compatibly to '<xyz>.this'.