FE: Disallow assigning '<xyz>.this' to type '<xyz>', fix #125
The problem is that '<xyz>.this' may be the type of a heir feature. This is ok
as long as it is a ref, but this will fail if the heir is a value type.  As long
as there is no Eiffel-style 'like this' mechanism in Fuzion, 'ref <xyz>' has to
be used for types that '<xyz>.this' should be assigned to.