AST/FE/ME/BE: Added AbstractFeature for features loaded from binary mir files
A step in a major restructuring to support modules that can be compiled to mir
files individually.  AbstractFeature shows the data required from a Feature
once it was stored in a MIR file.

Some places in the ME and BE still cst AbstractFeature to Feature, this code
needs to be improved to avoid the use of Feature altogether.