FE: Add LibraryFeature as an instance of AbstractFeature for precompiled code
In a precompiled library, features accesed from the AST will eventually be
accessed through instances of LibraryFeature.  For now, LibraryFeature justs
wraps around a ast.Feature and passes the information through.

In a next step. LibraryFeature will become independent from Feature and access
binary code generated by the front end instead of the AST. This will be stepwise
extended until the AST of the module is no longer needed.

Until this is done, we have the somewhat ugly situation that features in a
module exist twice, once in the AST and as an instance of LibraryFeature.  This
makes it impossible to, e.g., compare Features using '==' if one is not sure
which of these two versions we have.  There is a new method
AbstractFeature.sameAs for this comparison. Additionaly, methods astFeature and
libraryFeature provide the specific instance if it is available.