AST/FE: Implement LibraryFeature.arguments() using mir file data
Inner features written to a mir file are now sorted: First, the formal arguments
are written in the order they appear in the feature declaration, then the
remaining features are written in whatever order FeatureName.compareTo uses.

This required a workaround in formal argument resolution to make sure
select(res, int) is called on ast.Feature.  Eventually, the mechanism that
creates a new Feature for every select-ed element needs to be changed since we
cannot write these features out to a library file.