FE/AIR: Moved Handling of select for open generic fields from FE to ME/AIR
Before this patch, for open generics such a tuple.values, one feature was
generated for each generic variant.  This does not work with libraries since the
number of variants is no longer known. So this patch moves the creation of the
variants to the middle end: For each open generic field, there is only one
Feature, but there might be arbitrarily many Clazz'es.

This avoids the need for AbstractFeature.select and related methods, but it
requires Call to carry its _select field into the .mir files.