ast/fe/me/mir: Replaced Module.redefinitions* by AbstractFeature.redefines
The set of redefinitions can grow arbitrarily, while the set of redefined
features is known for each feature.  So collecting the latter will permit that
set to be saved for a module file.

During findUsedFeatures phase, the set of redefinitions is required. This is now
created dynamically during that phase, containing only those features that have
been found to be used.