air/ast/be: Removed Box._stmnt and Box._arg fields and related code
These fields are not set for Box loaded from a library resulting in errors
during execution since Box does not work properly in all cases.

Instead of Box dragging along a reference to the statement that created the Box,
now findClazzes phase now propagates the expected clazz from statements that use
the value to the Box statements in Clazzes.propagateExpectedClazz.