ast/fe/C backend: Replaced AbstractFeature.compareTo
The previous code was somewhat a hack using System.identityHashCode which has
the main disadvantage that the order defined by compareTo() may very between
program runs.

Now, an counter is used to equip instances of Feature with a unique id that can
be compared.  LibraryFeatures already have such an id in form of their index. It
remains to define the order between Feature and LibraryFeature.

Since the order of features now corresponds to the source code order instead of
alphabetical order comparing feature names, the ids used in C field names
created by the C backend had to be fixed in the initializaion of string
constants (now field 'length' as an argument is before 'data', opposite to the
alphabetical order).