be/interpreter: fix java interface's intrinsics fuzion.java.call(V|S|C)0
The problem was that these expect a 'sys.array<JavaObject>' as last argument,
while this argu is used as if it was 'sys.array<JavaObject>.data'.  This worked
only by accident since 'sys.array' is a value type, while 'data' is also a value
type and used to be stored at offset 0.

The recent changes in field comparison results in the offset to be 1 such that
we access undefind data.