parser/ast/fe: Join 'unit' and 'stmntsEof' into one parser rule 'unit', fix #136
This avoids complexity of having two different main nodes of the AST depending
on whether we parse a normal file or we parse stdin (or a file given on the
command line, which is treated equally).  Now, the AST always starts with a list
of Stmnts produced by 'unit'.

These statements may be feature declarations that will get added to the
universe, or Stmnts that are executable. Executable statements are allowed only
if read from stdin (or a file provided directly)) and will be added to the
universe.

This also fixes the slight syntactic difference shown in issue #136.