tests: Added small test of Java interface
I have overlooked that an earlier patch broke the Java interface, so I better
have a test for it.

This currently does not work for the C backend (since there is no Java interface
for the C backend yet), and it fails when PREcONDITIONS or POSTCONDITIONS are
enabled.  Must be checked why this is the case.