ast/fe: Proper support for defining features with qualified feature name
This allows adding features to existing library features, e.g.

  i32.factors => 1..val & x -> val %% x
  i32.postfix ! => if val <= 1 then 1 else val * (val-1)!

Which would then allow to write '42.factors' or '[x,y,z].length!' to get the
list of 42's factors or calculate the factorial of an array length.

It is no longer allowed to declare a fully qualified field such as

  a.b.c.f := 3

The reason is that it is unclear when the initialization code should be
executed.

The error message '...is declared in wrong environmant' no longer exists.
Instead, when lookup of an outer feature fails, the corresponging error will
report that that type was not be found.