air/fuir/C backend: Avoid hard-coded field names for const strings
This helps for issue #161: Instead of using hard-coded names for fields, get the
corresponding clazz'es from FUIR and run them through the C backends name
mangling which uses the index that FUIR assigned to these fields.

For this, had to ad fields to Clazzes.java for clazzes
'conststring.internalArray', 'sys.arrray<u8>' and its fields 'data' and
'length'.

Renamed 'Clazzes.constStringBytesArray' as 'Clazzes.sysArray_u8' since this is
the clazz for type 'sys.array<u8>', which is also the result type of
'conststring.internalArray'.