lib: Do not return 'outer.this' as a value type, fixes first example in #166
Returning 'outer.this' in a feature returning a value type does not work for
heirs for 'outer' since then, 'outer.this' is a different, incompatible value
type or maybe even a 'ref' type.  In both cases, the interpreter worked
more-or-less by accident, while the C backend reports a fully justified type
error.