air: Add new category of clazz: dynamic outer ref, prerequisite for #176 & #30
The new category is for clazzes whose instances are value types and who are
referenced from inner instances and have heirs such that dynamic binding is
required for the references from these inner clazzes.

In a next step, type information will need to be added for dynamic binding to
work.

This patch also does some cleanup in the middle end building the IR. It added
Clazz.parents() to get a set of all parents, similar to Clazz.heirs().