air: Fix failure of tests/inheritance with C backend
The problem was that a feature y that was called in an inheritance call and
directly was not marked as called due to caching of the clazz created for the
inheritance call and hence no code was created.

Now, the code to set that a feature is called and instantiated is executed even
for cached clazzes.