ast: Make sure features of internally used types are resolved for types early.
Fixes #209. The problem was that an error for a mismatch in the generics used in
a parent call of internally used types such as 'i32' was reported later than the
preconditions checked if that error was actually reported.

Forcing early resolution for types avoids this problem, but it might still be
possible to change the code of internally used features auch as 'i32' or their
parent such that similar problem occur.

Also weakened precondition in one constructor of Type.