lib: onewayMonad now can also be used for normal monads
A new parameter passed to onewayMonad decides if this is to be used as a normal
'plain' monad, or if a onewayMonad is to be 'inst'alled while calculating a
function or if a onewayMonad that is installed needs to be 'repl'aced.

This allows the 'handles' monad to be a heir of 'onewayMonad' in a way that
'handles' can be used as a plain monad or as a oneway monad.

Added features to handles and handles2 such that they can be used when installed
in the enviroment as a oneway monad.

Also changes 'onewayHandles' and 'onewayHandles2' accordingly, even though
these are not subject to be removed since there 'plain' counterparts are now usable as oneway monads.