lib/interpreter: First experimental support for onewayMonad.abort
For the interpreter backend, 'abort' performs a longjump to the installation
of a onewayMonad to return immediately.

For the C backend, only dummy code is created for 'abortable' and 'abort',
'abort' stops the code with an error while 'abortable' just calles the
provided function.

During findClazzes phase, a new method 'Clazz.intrinsicCalled' now takes
care to add all clazzes required by intrinsics, in particular the clazzes
returnd by intrinsic_constructor|s and the clazz of the 'call' that
is performed by 'abortable'.