C backend: Fix #168 for non-constructors
The factors11.fz and factors12.fz examples from flang.dev did not longer work
with the fix for #215 since here. the routine 'i32.infix .. .infix &' is not a
constructor, but nevertheless causes the outer value instance to survive the
call to 'infix ..'.

This patch removes the restriction to constructors for explicit heap allocation
or outer value instances.