parser: Allow constructor declaration without code to end with fullStop ('.').
This is a little tricky since '.' is used in a call as well.  The white space is
important: A full stop is a '.' that follows directly the previous token and
that is followed by white space. This means that

  a.
  b.
  c.

declares three constructors a, b, and c while

  a.b.c

calls 'a.b.c'