parser: Add syntax sugar for easier enum declaration using 'of'
The idea is that previously, a declaration of a choice had the form

  red is
  green is
  blue is
  point(x,y i32) is

  colorOrPoint : choice<red, green, blue, point> is

Now, this can be done using 'of' as follows

  colorOrPoint : choice of
    red is
    green is
    blue is
    point(x,y i32) is

or even

  colorOrPoint : choice of
    red, green, blue.
    point(x,y i32).

The semantics is: All decarations after 'of' will be added as inner features to
the outer feature (of colorOrPoint in this example). The type parameters of the
last parent will be extended by the corresponding types, read, green, blue and
point, in this case.