parser, FE: Alloc type parameters in the formal arguments list
Type parameters are now argument features and inherited etc. just as other
features.

Instead of writing

  orderedMap<K : ordered<K>, V>(ks array<K>,
                                vs array<V>) : map<K, V>

one can now write

  orderedMap(OK ordered<OK>.type,
             redef V type,
             ks array<OK>,
             vs array<V>) : map<OK, V>

Note that we must avoid name clashes with inherited type parameters (using 'OK'
instead of 'K' in the example) or explicitly 'redef' inherited type parameters
(as for 'V' in the example).

The old syntax is still supported but will be removed. Internally, all type
parameters defined the old way using '<' and '>' are converted to features of
type parameter.  The FormalGenerics instance is then created from these type
parameters.

Eventually, TypeArgs will be removed from .fum file and type parameters will be
saved to .fum files like other features, but this is not the case yet.

Open type parameters as in

   Function<R,A...> ref is

can be written as

   Function(R type, A type...) ref is

Where renaming or 'redef' annotations were required, the stdlib has been changed
to the new syntax.