lib/C + Interpreter backend: Add effect 'concur.thread' with operation 'spawn'
'spawn' receives a function as an argument and will calculate the value of that
function in parallel.

New threads do not inherit the effects environment of the existing thread, but
they can access the closure.

This is purely experimental, no synchornization mechanisms nor inter-thread
communication primitives are supported at this point. Also, the control flow
graph analysis does not work yet for spawn'ed code.

The interpreter backend now stores thread local data in an instance of
FuzionThread that is created for each thread.

The C backend uses static fields make thread local using '__thread'.