parser: Type now allow whitespace and parentheses to separate type parameters
So instead of writing

  q(l list<map<string, i32>>) unit is
    for m in l do ...

we can now write

  z(l list(map string i32)) unit is
    for m in l do

So we are diverting from common syntax in favor of something shorter and
clearer.

Also, this will solve some parsing issues once the notation using '<' and '>' is
abandoned since the lexer creates operator tokens like '>>', that have to be
split into two '>' in case they end a list of type parameters.