lib/fe/me/be: Add support for type parameters is .fum file and in calls
In a feature 'f(X type)' with a type parameter 'X' it is now possible to call
'X' as in 'say X'.

The result of calling a type parameter is an instance of feature 'Type' that
provides 'asString' and 'name' features. The outer instance is the the type's
outer types 'Type' instance.

Currently, the result is a ref type to an instance containing one field that
holds the name.  This should be replaced by a unit type value with no state such
that no data needs to be copied.