AIR: Add support to inherit feature with open generic result type, fix #292
An feature with an open generic result type that was inherited was not
replicated properly for the number of actual type parameters.

In the example of #292, this resulted in a cast added to the C code applied to
an i32 call argument while the cast should have been added to the call
target. The wrong value was used since the call argument count got lost at this
particular location since the open generics where replaced by an empty list
instead of the actual list of type parameters.