AST: Make 'result' visible in post-condition of routine using '=>', fix#45
Also had to change the order when visiting features: contract is visited after
the implementation since the implementation determines the type of the result,
which is then used in the post condition.

'result' is now also visible in pre-condition and in the implementation, but
reading it will result in reading an uninitialized field.