fuir: Add methods to navigate back and forth withing FUIR's code blocks
'codeIndex(c,ix,delta)' can jump back or forward a number of expressions
relative to 'ix'.

'skipBack' can jump backward all the expressions that are part of the creation
of the value on the top of the stack.

CAVEAT: Going backward with these functions easily results in quadratic
performance since they need to iterate over all expressions starting at index 0.