be/C: Add tail recursion optimization, fix #133
When a call is found to be a tail call and one of the targets called is the
current clazz, the call will be done as a goto to the start of the current
clazz' code.

In a tail recursive call, any arguments passed to the call will be assigned to
the arguments variables of the currently created C function.

Issue #236 still remains since this uses a tail recursion with a different
target value, which is not supported yet.